backend/requirements.txt
asgiref==3.5.2
Django==3.2.15
django-colorfield==0.7.2
django-debug-toolbar==3.2.4
django-filter==22.1
djangorestframework==3.14.0
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
drf-extra-fields==3.4.0
gunicorn==20.1.0
psycopg2-binary==2.9.6
PyJWT==2.5.0
python-dotenv==0.21.0
pytz==2022.2.1
reportlab==3.6.11
requests==2.28.1
sqlparse==0.4.3